proper color patern in the comments
