+++ app/src/main/res/layout/fragment_comments.xml
@@ -10,9 +10,8 @@
     android:layout_height="match_parent">
 
     <android.support.v7.widget.RecyclerView
-      android:clipToPadding="false"
+      style="@style/HNews.List"
       android:id="@+id/list_comments"
-      android:padding="@dimen/comment_list_padding"
       android:layout_width="match_parent"
       android:layout_height="match_parent" />
