+++ app/src/main/res/layout/view_comment.xml
@@ -2,6 +2,7 @@
 
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
+  android:background="@color/white"
   android:id="@+id/comment_root"
   android:layout_width="match_parent"
   android:layout_height="wrap_content"
@@ -10,13 +11,13 @@
   <View
     android:layout_width="match_parent"
     android:layout_height="1dp"
-    android:layout_marginTop="@dimen/card_content_padding"
     android:background="@drawable/divider_horizontal_bright" />
 
   <LinearLayout
     android:layout_width="match_parent"
     android:layout_height="@dimen/default_item_height"
-    android:orientation="horizontal">
+    android:orientation="horizontal"
+    android:padding="@dimen/item_content_padding">
 
     <com.malmstein.yahnac.views.fonts.YahnacTextView
       style="@style/HNews.CommentHeader"
@@ -50,6 +51,8 @@
     android:id="@+id/comment_text"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:paddingRight="@dimen/item_content_padding"
+    android:paddingLeft="@dimen/item_content_padding"
     android:paddingBottom="@dimen/content_bottom_padding"
     tools:text="This is a comment" />
