+++ app/src/main/res/layout/view_story_header.xml
